Removed time zone
This led to a nasty bug: As the saldomat feed sets 22:00 as time for his items this is transferred to the created_at attribute of the payment.
If two hours are added for e.g. Berlin time zone this means that this payment is moved to the next day or even the next month (at the last day of a month)